software/laptop/requirements.txt
paho-mqtt
numpy
scipy
